+++ src/Eyewitness/Eye/Commands/InstallCommand.php
@@ -44,6 +44,8 @@ class InstallCommand extends Command
             return;
         }
 
+        $this->displayInitialInstallMessage();
+
         $this->call('config:publish', ['package' => 'eyewitness/eye4']);
 
         try {
@@ -53,7 +55,7 @@ class InstallCommand extends Command
         }
 
         try {
-            $app = $this->getApplicationData();
+            $app = $this->eye->runAllChecks(false);
         } catch (Exception $e) {
             return $this->failedInstallation('application data', $e->getMessage());
         }
@@ -102,17 +104,16 @@ class InstallCommand extends Command
     }
 
     /**
-     * Get all the relevant application data.
+     * Display the initial installation message.
      *
-     * @return array
+     * @return void
      */
-    protected function getApplicationData()
+    protected function displayInitialInstallMessage()
     {
-        $app = $this->eye->runAllChecks(false);
-        $app['name'] = $this->handleAppName('My Laravel App');
-        $app['domain'] = $this->handleAppDomain(Config::get('app.url'));
-
-        return $app;
+        $this->info('______________________________________');
+        $this->info(' ');
+        $this->info('Installing and configuring package for Eyewitness.io....');
+        $this->info(' ');
     }
 
     /**
@@ -122,44 +123,17 @@ class InstallCommand extends Command
      */
     protected function displayOutcome()
     {
-        $this->info('-------------------------------');
+        $this->info('______________________________________');
+        $this->info(' ');
         $this->info('Success! Your Laravel application has been configured and is now being monitored by Eyewitness.io!');
         $this->info(' ');
-        $this->info('App Token: '.Config::get('eye::app_token'));
-        $this->info('Secret Key: '.Config::get('eye::secret_key'));
+        $this->info('   App Token: '.config('eyewitness.app_token'));
+        $this->info('   Secret Key: '.config('eyewitness.secret_key'));
         $this->info(' ');
         $this->info('Optional: You can provide your email address now, and we will send you an email with the "app_token" and "secret_key" included. This email will be sent by our email server, so it is ok if you do not have email configured on this server. No spam, we promise!');
         $this->info(' ');
 
         $this->handleOptionalEmailRequest();
-    }
-
-    /**
-     * Handle confirming the application domain.
-     *
-     * @param  string   $url
-     * @return string
-     */
-    protected function handleAppDomain($url)
-    {
-        return $this->ask('2. What is the full domain of your application that we are monitoring (so we can ping the server)?', $url);
-    }
-
-    /**
-     * Handle confirming the application name.
-     *
-     * @param  string   $name
-     * @return string
-     */
-    protected function handleAppName($name)
-    {
-        $this->info('-------------------------------');
-        $this->info(' ');
-        $this->info('Installing and configuring package for Eyewitness.io.');
-        $this->info(' ');
-        $this->info('There are just two questions to help get you started.');
-        $this->info(' ');
-        return $this->ask('1. Please give your application a name (this helps when you are monitoring multiple applications):', $name);
     }
 
     /**
@@ -173,22 +147,23 @@ class InstallCommand extends Command
             $email = $this->ask('Your email address (optional - leave blank if you do not want an email):', false);
 
             if ($email === false) {
+                $this->info('______________________________________');
                 $this->info(' ');
                 $this->info('Now that your package is installed - please head to https://eyewitness.io and login to view your application monitor. You will need your "app_token" and "secret_key" to access your server on the website.');
                 $this->info(' ');
                 $this->info('You can copy and paste these from above. Or you can get them from your '.app_path('/config/packages/eyewitness/eye/config.php').' file as well.');
-                break;
-            } else {
-                if ($this->validateEmail($email)) {
-                    $this->eye->api()->sendInstallEmail($email);
-                    $this->info(' ');
-                    $this->info('Email sent via Eyewitness.io to "'.$email.'". Please check your inbox for a copy of your "app_token" and "secret_key"');
-                    $this->info(' ');
-                    break;
-                } else {
-                    $this->error('Sorry - that email address does not seem to be valid. Please try again.');
-                }
+                return;
             }
+
+            if ($this->validateEmail($email)) {
+                $this->eye->api()->sendInstallEmail($email);
+                $this->info(' ');
+                $this->info('Email sent via Eyewitness.io to "'.$email.'". Please check your inbox for a copy of your "app_token" and "secret_key"');
+                $this->info(' ');
+                return;
+            }
+
+            $this->error('Sorry - that email address does not seem to be valid. Please try again.');
         }
     }